Update with cache and langauge model example
